backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt
gunicorn
whitenoise>=5.3.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
dj-database-url>=1.0
psycopg2-binary>=2.9.9
PyPDF2>=3.0.0
python-docx>=1.1.0
Pillow>=10.0.0
python-magic>=0.4.27